Adds a simple pagination to journeys
