+++ packages/client/src/App.tsx
@@ -4,9 +4,9 @@ import { Route, Routes, useParams } from 'react-router-dom';
 import Stations from './pages/Stations';
 import Journeys from './pages/Journeys';
 import Navbar from './pages/Navbar';
-import { CityBikesJourney, CityBikeStation } from './table/types';
+import { CityBikeStation } from './table/types';
 
-const PAGE_SIZE = 100
+export const PAGE_SIZE = 20
 
 const SingleStaionView = () => {
   const [stationDetails, setStationDetails] = useState<CityBikeStation>()
@@ -39,7 +39,6 @@ const SingleStaionView = () => {
 
 const App = () => {
   const [stations, setStations] = useState<CityBikeStation[]>([])
-  const [journeys, setJourneys] = useState<CityBikesJourney[]>([])
 
   const fetchStations = async () => {
     const response = await fetch(`/stations?pageSize=${PAGE_SIZE}`)
@@ -47,17 +46,8 @@ const App = () => {
 
     setStations(data)
   }
-
-  const fetchjourneys = async () => {
-    const response = await fetch(`/journeys?pageSize=${PAGE_SIZE}`)
-    const data = await response.json()
-
-    setJourneys(data)
-  }
-
   useEffect(() => {
     void fetchStations()
-    void fetchjourneys()
   }, [])
 
   return (
@@ -65,7 +55,7 @@ const App = () => {
       <Navbar />
       <Routes>
         <Route path='/' element={<Stations data={stations} />} />
-        <Route path='/journeys' element={<Journeys data={journeys} />} />
+        <Route path='/journeys' element={<Journeys />} />
         <Route path='/station/:stationId' element={<SingleStaionView />} />
       </Routes>
     </>
