+++ packages/client/src/pages/Journeys.tsx
@@ -1,5 +1,14 @@
+import { useState, useEffect } from "react"
+import styled from "styled-components"
+import { PAGE_SIZE } from "../App"
 import Table from "../table/Table"
 import { ColumnDefinitionType, CityBikesJourney } from "../table/types"
+
+const ButtonWrapper = styled.div`
+    display:flex;
+    align-items: center;
+    justify-content: center;
+`
 
 const columns: ColumnDefinitionType<CityBikesJourney, keyof CityBikesJourney>[] = [
     {
@@ -21,12 +30,22 @@ const columns: ColumnDefinitionType<CityBikesJourney, keyof CityBikesJourney>[] 
     }
 ]
 
-interface JourneysProps {
-    data: CityBikesJourney[]
-}
+const Journeys = () => {
+    const [journeys, setJourneys] = useState<CityBikesJourney[]>([])
+    const [page, setPage] = useState(1)
 
-const Journeys = ({ data }: JourneysProps) => {
-    const mappedData = data.map((journey) => {
+    useEffect(() => {
+        const fetchjourneys = async () => {
+            const response = await fetch(`/journeys?pageSize=${PAGE_SIZE}&page=${page}`)
+            const data = await response.json()
+
+            setJourneys(data)
+        }
+
+        void fetchjourneys()
+    }, [page])
+
+    const mappedData = journeys.map((journey) => {
         return {
             ...journey,
             covered_distance_in_meters: `${(Number(journey.covered_distance_in_meters) / 1000).toFixed(1)}`,
@@ -38,6 +57,11 @@ const Journeys = ({ data }: JourneysProps) => {
         <>
             <p> Journey list</p>
             <Table data={mappedData} columns={columns} />
+
+            <ButtonWrapper>
+                <button disabled={page === 1} onClick={() => setPage((prevState) => prevState - 1)}>Prev page</button>
+                <button onClick={() => setPage((prevState) => prevState + 1)}>Next page</button>
+            </ButtonWrapper>
         </>
     )
 }
